Gstreamer 1.0 support

+++ audio/player/recorder.py
@@ -7,30 +7,38 @@ import sys
 import time
 import datetime
 
-import gobject
+import gi
+gi.require_version('Gst', '1.0')
+from gi.repository import Gst, GLib
 
-gobject.threads_init()
-import pygst
+if getattr(sys, 'frozen', False):
+    basedir = sys._MEIPASS
+    # Gst.debug_set_default_threshold(4)
+    # Gst.debug_set_active(True)
+    # Gst.segtrap_set_enabled(False)
+    # Gst.Registry.fork_set_enabled(False)
+# else:
+    # dir = os.path.dirname(__file__)
+    # basedir = os.path.join(dir, 'dist', 'audio')
+    # Gst.debug_set_default_threshold(3)
+    # Gst.debug_set_active(True)
+    # Gst.Registry.fork_set_enabled(False)
+    # os.environ['GST_PLUGIN_PATH'] = os.path.join(basedir, 'gst_plugins')
+    # basedir = os.path.dirname(__file__)
 
-pygst.require('0.10')
-import gst
+Gst.init(None)
 
 from PyQt4 import QtCore
 
 from audio.core import Registry, Settings, Utils
 
-if getattr(sys, 'frozen', False):
-    basedir = sys._MEIPASS
-else:
-    basedir = os.path.dirname(__file__)
-
 
 class Recorder(QtCore.QThread):
-    updatemeter = QtCore.pyqtSignal(gst.Message)
+    updatemeter = QtCore.pyqtSignal(object)
 
     def __init__(self):
         QtCore.QThread.__init__(self)
-        self.loop = gobject.MainLoop()
+        self.loop = GLib.MainLoop.new(None, False)
         self.filepath = None
         self.pipelineactive = False
         self.recording = False
@@ -39,68 +47,75 @@ class Recorder(QtCore.QThread):
 
         self.settings = Settings()
         self.utils = Utils()
-
-        self.pipeline = gst.Pipeline("Recording Pipeline")
-
+        self.device_monitor = Gst.DeviceMonitor()
+        print(self.device_monitor.get_devices())
+        self.pipeline = Gst.Pipeline()
         if sys.platform == 'darwin':
-            self.audiosrc = gst.element_factory_make("osxaudiosrc", "audiosrc")
-            self.srcratecap = gst.Caps("audio/x-raw-float, rate=" + self.srcrate)
+            self.audiosrc = Gst.ElementFactory.make('osxaudiosrc')
+            self.srcratecap = Gst.Caps.from_string("audio/x-raw, rate=" + self.srcrate)
 
         elif os.name == 'nt':
-            self.audiosrc = gst.element_factory_make("dshowaudiosrc", "audiosrc")
-            self.srcratecap = gst.caps_new_any()
+            self.audiosrc = Gst.ElementFactory.make("dshowaudiosrc", "audiosrc")
+            self.srcratecap = Gst.Caps.new_any()
 
         else:
-            self.audiosrc = gst.element_factory_make("autoaudiosrc", "audiosrc")
-            self.srcratecap = gst.caps_new_any()
-
-        self.srcratefilter = gst.element_factory_make("capsfilter", "srcratefilter")
+            self.audiosrc = Gst.ElementFactory.make("autoaudiosrc", "audiosrc")
+            self.srcratecap = Gst.Caps.new_any()
+        self.srcratefilter = Gst.ElementFactory.make("capsfilter", "srcratefilter")
         self.srcratefilter.set_property("caps", self.srcratecap)
 
-        self.audioconvert = gst.element_factory_make("audioconvert", "audioconvert")
+        self.audioconvert = Gst.ElementFactory.make("audioconvert", "audioconvert")
 
-        self.audioresample = gst.element_factory_make("audioresample", "audioresample")
+        self.audioresample = Gst.ElementFactory.make("audioresample", "audioresample")
 
-        self.level = gst.element_factory_make("level", "level")
+        self.level = Gst.ElementFactory.make("level", "level")
 
-        self.recordingratecap = gst.caps_new_any()
-        self.recordingratefilter = gst.element_factory_make("capsfilter", "recordingratefilter")
+        self.recordingratecap = Gst.Caps.new_any()
+        self.recordingratefilter = Gst.ElementFactory.make("capsfilter", "recordingratefilter")
         self.recordingratefilter.set_property("caps", self.recordingratecap)
 
-        self.valve = gst.element_factory_make("valve", "valve")
+        self.valve = Gst.ElementFactory.make("valve", "valve")
 
-        self.sink = gst.Bin("Sink")
+        self.sink = Gst.Bin()
 
-        self.audiorate = gst.element_factory_make("audiorate", "audiorate")
+        self.audiorate = Gst.ElementFactory.make("audiorate", "audiorate")
         self.audiorate.set_property("skip-to-first", True)
-        self.wavenc = gst.element_factory_make("wavenc", "wavenc")
-        self.filesink = gst.element_factory_make("filesink", "filesink")
+        self.wavenc = Gst.ElementFactory.make("wavenc", "wavenc")
+        self.filesink = Gst.ElementFactory.make("filesink", "filesink")
 
-        self.sink.add_many(self.audiorate, self.wavenc)
+        self.sink.add(self.audiorate)
+        self.sink.add(self.wavenc)
         self.audiorate.link(self.wavenc)
 
         self.sinkpad = self.audiorate.get_static_pad("sink")
-        self.sinkghostpad = gst.GhostPad("sink", self.sinkpad)
+        self.sinkghostpad = Gst.GhostPad.new("sink", self.sinkpad)
         self.sinkghostpad.set_active(True)
         self.sink.add_pad(self.sinkghostpad)
 
-        if not (self.pipeline and self.audiosrc and self.srcratecap and self.srcratefilter and self.audioconvert and
-                self.audioresample and self.level and self.recordingratecap and self.recordingratefilter and
-                self.valve):
+        if not (self.pipeline and self.audiosrc and self.srcratefilter and self.audioconvert and self.audioresample
+                and self.level and self.recordingratefilter and self.valve):
             print("Not all elements could be loaded", sys.stderr)
             exit(-1)
 
-        self.pipeline.add(self.audiosrc, self.srcratefilter, self.audioconvert, self.audioresample, self.level,
-                          self.recordingratefilter, self.valve, self.sink)
+        self.pipeline.add(self.audiosrc)
+        self.pipeline.add(self.srcratefilter)
+        self.pipeline.add(self.audioconvert)
+        self.pipeline.add(self.audioresample)
+        self.pipeline.add(self.level)
+        self.pipeline.add(self.recordingratefilter)
+        self.pipeline.add(self.valve)
+        self.pipeline.add(self.sink)
 
-        if not gst.element_link_many(self.audiosrc, self.srcratefilter, self.audioconvert, self.audioresample,
-                                     self.level, self.recordingratefilter, self.valve, self.sink):
+        if not (self.audiosrc.link(self.srcratefilter), self.srcratefilter.link(self.audioconvert),
+                self.audioconvert.link(self.audioresample), self.audioresample.link(self.level),
+                self.level.link(self.recordingratefilter), self.recordingratefilter.link(self.valve),
+                self.valve.link(self.sink)):
             print("Elements could not be linked", sys.stderr)
             exit(-1)
 
         self.valve.set_property("drop", True)
 
-        self.pipeline.set_state(gst.STATE_PLAYING)
+        self.pipeline.set_state(Gst.State.PLAYING)
         self.pipelineactive = True
 
         self.bus = self.pipeline.get_bus()
@@ -113,24 +128,30 @@ class Recorder(QtCore.QThread):
         self.loop.run()
 
     def record(self):
-        self.load()
-        self.sink.add(self.filesink)
-        self.wavenc.link(self.filesink)
-        self.sink.set_state(gst.STATE_PLAYING)
-        self.valve.set_property("drop", False)
-        self.recording = True
+        if not self.recording:
+            self.load()
+            self.sink.add(self.filesink)
+            self.wavenc.link(self.filesink)
+            self.sink.set_state(Gst.State.PLAYING)
+            self.valve.set_property("drop", False)
+            self.recording = True
+        else:
+            pass
 
     def pause(self):
         self.valve.set_property("drop", True)
-        self.sink.set_state(gst.STATE_PAUSED)
+        self.sink.set_state(Gst.State.PAUSED)
 
     def stop(self):
-        self.valve.set_property("drop", True)
-        self.sink.send_event(gst.event_new_eos())
-        self.sink.set_state(gst.STATE_NULL)
-        self.wavenc.unlink(self.filesink)
-        self.sink.remove(self.filesink)
-        self.recording = False
+        if self.recording:
+            self.valve.set_property("drop", True)
+            self.sink.send_event(Gst.Event.new_eos())
+            self.sink.set_state(Gst.State.NULL)
+            self.wavenc.unlink(self.filesink)
+            self.sink.remove(self.filesink)
+            self.recording = False
+        else:
+            pass
 
     def on_message(self, bus, message):
         """
@@ -139,19 +160,16 @@ class Recorder(QtCore.QThread):
         :param message:
         """
         t = message.type
-        if t == gst.MESSAGE_EOS:
-            print("EOS")
-            self.pipeline.set_state(gst.STATE_NULL)
-            self.pipelineactive = False
+        if t == Gst.MessageType.EOS:
+            self.sink.set_state(Gst.State.NULL)
+            self.pipelineactive = True
             self.recording = False
-        elif t == gst.MESSAGE_ERROR:
-            self.pipeline.set_state(gst.STATE_NULL)
-            err, debug = message.parse_error()
-            print("Error: {0}".format(err), debug)
+        elif t == Gst.MessageType.ERROR:
+            self.pipeline.set_state(Gst.State.NULL)
             self.pipelineactive = False
             self.recording = False
         elif message.src == self.level:
-            self.updatemeter.emit(message)
+            self.updatemeter.emit(message.get_structure().get_value("rms"))
 
     def load(self):
         now = datetime.datetime.now()
@@ -159,7 +177,7 @@ class Recorder(QtCore.QThread):
 
         if not self.recording:
             self.recordrate = self.settings.value("RecordingSampleRate")
-            self.recordingratecap = gst.Caps("audio/x-raw-int, rate=" + self.recordrate)
+            self.recordingratecap = Gst.caps_from_string("audio/x-raw, rate=" + self.recordrate)
             self.recordingratefilter.set_property("caps", self.recordingratecap)
             self.filepath = os.path.join(self.settings.value("RecordingDirectory"), self.utils.clean_name(date))
             self.filesink.set_property("location", self.filepath)
@@ -167,6 +185,6 @@ class Recorder(QtCore.QThread):
             pass
 
     def stop_loop(self):
-        self.pipeline.send_event(gst.event_new_eos())
-        self.pipeline.set_state(gst.STATE_NULL)
+        self.pipeline.send_event(Gst.Event.new_eos())
+        self.pipeline.set_state(Gst.State.NULL)
         self.loop.quit()
